move create-snapshot command to test command
