+++ nf_core/modules/__init__.py
@@ -6,7 +6,6 @@ from .lint import ModuleLint
 from .list import ModuleList
 from .modules_json import ModulesJson
 from .modules_repo import ModulesRepo
-from .modules_test import ModulesTest
 from .modules_utils import ModuleException
 from .patch import ModulePatch
 from .remove import ModuleRemove
