+++ nf_core/subworkflows/__init__.py
@@ -4,5 +4,4 @@ from .install import SubworkflowInstall
 from .lint import SubworkflowLint
 from .list import SubworkflowList
 from .remove import SubworkflowRemove
-from .subworkflows_test import SubworkflowsTest
 from .update import SubworkflowUpdate
